+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Add glossary management support.
+
+
 ## [0.4.1] - 2022-08-12
 ### Changed
 * Update contributing guidelines, we can now accept Pull Requests.
@@ -44,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v0.4.1...HEAD
 [0.4.1]: https://github.com/DeepLcom/deepl-php/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/DeepLcom/deepl-php/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/DeepLcom/deepl-php/compare/v0.2.0...v0.3.0